Renamed Types.tsx to Row.tsx; Renamed based on the file content / prevent ambiguity.
